refactor: my_social_list_card_item 하나로 통합
